+++ packages/nimbus/configs/babel.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 const { getConfig } = require('@rajzik/config-babel');
-const { getSettings } = require('@rajzik/nimbus-common');
+const { getSettings } = require('@rajzik/lumos-common');
 
 const { context, tool } = process.beemo;
 const { graphql, library, next, node, react, env } = getSettings();